services/simulator/requirements.txt
fastapi
uvicorn
prometheus_client
pytest
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp